+++ src/finalizer/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./polygon";
+export * from "./arbitrum";
